requirements.txt
langchain
langchain-astradb
langchain-huggingface
langchain-groq
langchain-community
datasets
pypdf
python-dotenv
pandas
flasks
prometheus_client